- implemented execution of opcode 2NNN - adjusted call stack data type

+++ src/main/java/emulator/CallStack.java
@@ -1,6 +1,6 @@
 /*
  * Developed by Patrick Metz <patrickmetz@web.de>.
- * Last modified 25.02.19 18:37.
+ * Last modified 26.02.19 13:41.
  * Copyright (c) 2019. All rights reserved.
  */
 
@@ -10,17 +10,17 @@ import java.util.Stack;
 
 final class CallStack {
 
-    private final Stack<Byte> memory;
+    private final Stack<Short> memory;
 
     CallStack() {
         memory = new Stack<>();
     }
 
-    byte pop() {
+    Short pop() {
         return memory.pop();
     }
 
-    void push(byte value) {
+    void push(Short value) {
         memory.push(value);
     }
